requirements-global.txt
# base requirements
numpy==2.2.2
opencv-python==4.11.0.86
pillow==11.1.0
tqdm==4.67.1
PyYAML==6.0.2
huggingface-hub==0.28.1
scipy==1.15.1; sys_platform != 'win32'
matplotlib==3.10.0
av==14.1.0 # using an older version. only update once torchvision has been updated as well.

# pytorch
accelerate==1.3.0
safetensors==0.5.2
tensorboard==2.18.0
pytorch-lightning==2.5.0.post0

# stable diffusion
-e git+https://github.com/huggingface/diffusers.git@464374f#egg=diffusers
transformers==4.48.3
sentencepiece==0.2.0 # transitive dependency of transformers for tokenizer loading
omegaconf==2.3.0 # needed to load stable diffusion from single ckpt files
invisible-watermark==0.2.0 # needed for the SDXL pipeline

# other models
pooch==1.8.2
open-clip-torch==2.30.0

# data loader
-e git+https://github.com/declangibbons23/mgds.git@dd95fa6#egg=mgds

# optimizers
dadaptation==3.2 # dadaptation optimizers
lion-pytorch==0.2.3 # lion optimizer
prodigyopt==1.1.2 # prodigy optimizer
schedulefree==1.4.0 # schedule-free optimizers
pytorch_optimizer==3.4.0 # pytorch optimizers

# Profiling
scalene==1.5.51

# ui
customtkinter==5.2.2

# cloud
runpod==1.7.7
fabric==3.2.2
